rename files + remove getConfigOrThrow
